[IMP] Add unit tests to cover the following use case scenario.
Steps to reproduce:

* create a new product 'Product Test': stockable, Invoicing Policy as Delivered quantities, and tracking by lots
* update its quantity to 3 with new lot 'Lot1'
* update its quantity to 3 with new lot 'Lot2'
* sale 4 of 'Product Test', confirm the SO
* on the delivery, split the lots and set To Do quantities as 3 for Lot1 and leave 0 for Lot2
* validate the delivery and create a backorder

Current behavior:

the field 'Production Lots' of the invoice generated is filled with both lots and, consequently,
the invoice report will also contain info related to the lot not delivered yet.

Expected behavior:

the invoice report only contains info related to the lot involved in the delivery, which is Lot1.